HARP-8067: Move the data adapters to sub directories.
This change moves the data adapters to their own sub directory.
