+++ @here/harp-vectortile-datasource/lib/adapters/omv/proto/vector_tile.proto
