+++ test/performance/OmvDecoderPerformanceTest.ts
@@ -26,13 +26,13 @@ import {
     OmvRestClient,
     OmvRestClientParameters
 } from "@here/harp-vectortile-datasource";
+import { OmvDataAdapter } from "@here/harp-vectortile-datasource/lib/adapters/omv/OmvDataAdapter";
 import { DecodeInfo } from "@here/harp-vectortile-datasource/lib/DecodeInfo";
 import {
     IGeometryProcessor,
     ILineGeometry,
     IPolygonGeometry
 } from "@here/harp-vectortile-datasource/lib/IGeometryProcessor";
-import { OmvProtobufDataAdapter } from "@here/harp-vectortile-datasource/lib/OmvData";
 import { OmvDecoder } from "@here/harp-vectortile-datasource/lib/OmvDecoder";
 
 export interface OMVDecoderPerformanceTestOptions {
@@ -141,7 +141,7 @@ export function createOMVDecoderPerformanceTest(
 
             await measurePerformanceSync(counterName, repeats, function() {
                 for (const [tileKey, tileData] of omvTiles) {
-                    const decoder = new OmvProtobufDataAdapter(geometryProcessor, undefined);
+                    const decoder = new OmvDataAdapter(geometryProcessor, undefined);
                     const decodeInfo = new DecodeInfo("profiler", mercatorProjection, tileKey, 0);
                     decoder.process(tileData, decodeInfo);
                 }
